Restructure and add unit test for drawing
