+++ lib/draw.js
@@ -5,13 +5,13 @@ var _ = require('lodash');
 
 var geopattern = require('geopattern');
 
-var fontSize = require('./lib/fontsize');
-var titleParts = require('./lib/titleparts');
-var svgCompile = require('./lib/svgcompile');
-var svgRender = require('./lib/svgrender');
+var fontSize = require('./fontsize');
+var titleParts = require('./titleparts');
+var svgCompile = require('./svgcompile');
+var svgRender = require('./svgrender');
 
-var topics = require('./lib/topic/');
-var colors = require('./lib/topic/colors.json');
+var topics = require('./topic/');
+var colors = require('./topic/colors.json');
 
 module.exports = function(output, options) {
 
@@ -27,8 +27,8 @@ module.exports = function(output, options) {
             "family": "Arial",
             "color": '#424242'
         },
-        "template": path.join(__dirname, "templates/default.svg"),
-        "watermark": path.join(__dirname, "published-with-gitbook.svg"),
+        "template": path.join(__dirname, "../templates/default.svg"),
+        "watermark": path.join(__dirname, "../templates/published-with-gitbook.svg"),
         "size": {
             "w": 1800,
             "h": 2360
