+++ lib/index.js
@@ -4,7 +4,7 @@ var _ = require('lodash');
 var Q = require('q');
 
 var createCover = require('./draw');
-var resize = require('./lib/resize');
+var resize = require('./resize');
 
 
 var copy = function(from, to) {
@@ -52,16 +52,23 @@ module.exports = {
                     );
                 }
 
+                console.log('createCover', path.join(outputDir, "cover.jpg"),
+                    _.extend({}, {
+                        title: that.options.title,
+                        author: that.options.author
+                    }, that.options.pluginsConfig.autocover))
                 return createCover(
-                path.join(outputDir, "cover.jpg"),
-                _.extend({}, {
-                    title: that.options.title,
-                    author: that.options.author
-                }, that.options.pluginsConfig.autocover));
+                    path.join(outputDir, "cover.jpg"),
+                    _.extend({}, {
+                        title: that.options.title,
+                        author: that.options.author
+                    }, that.options.pluginsConfig.autocover)
+                );
             })
 
             // Generate small cover
             .then(function() {
+                console.log('cover is ready');
                 // Check if a cover already exists in the output
                 if (fs.existsSync(path.join(outputDir, "cover_small.jpg"))) return Q();
